feat: 쿼리를 위한 서비스와 커맨드를 위한 서비스 분리
