+++ src/main/java/com/map/gaja/client/apllication/ClientQueryService.java
@@ -1,0 +1,33 @@
+package com.map.gaja.client.apllication;
+
+import com.map.gaja.client.domain.model.Client;
+import com.map.gaja.client.infrastructure.repository.ClientRepository;
+import com.map.gaja.client.presentation.dto.response.ClientListResponse;
+import com.map.gaja.client.presentation.dto.response.ClientResponse;
+import com.map.gaja.client.presentation.exception.ClientNotFoundException;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+import static com.map.gaja.client.apllication.ClientConvertor.*;
+
+@Service
+@RequiredArgsConstructor
+@Transactional(readOnly = true)
+public class ClientQueryService {
+    private final ClientRepository clientRepository;
+
+    public ClientResponse findUser(Long clientId) {
+        Client client = clientRepository.findById(clientId)
+                .orElseThrow(() -> new ClientNotFoundException(clientId));
+        return entityToDto(client);
+    }
+
+    public ClientListResponse findUser(String name) {
+        List<Client> clients = clientRepository.mockFindClientByCondition(name);
+        return entityToDto(clients);
+    }
+
+}
